Start using spire where possible.
One of the benefits of not being scala is that I don't have
to NIH everything. Spire is one of the finest scala libraries
there is and instead of continuing to reinvent it I would do
better to exploit it. My only hesitation is that it's a large
dependency, but one way or another we will bring that down to
the parts we actually depend upon.

For now we'll accept it in its entirety.

First up is

  type BooleanAlgebra[R] = spire.algebra.BooleanAlgebra[R]

And other than the naming choice of not vs. complement,
it works as-is. Although I see in the development version has
already changed the name from BooleanAlgebra to Bool. I guess
that's the sort of thing to which we will have to become
accustomed if we're going to have dependencies.